requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=0.0.1
langchain-text-splitters>=0.3.0
numpy>=1.26.0
tavily-python>=0.4.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
trafilatura>=1.8.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
fastapi>=0.110.0
uvicorn>=0.27.0
tenacity>=8.2.0
python-dotenv>=1.0.0
